doc: Add Points of Interest document
This will be one location where I unify some documentation related to
how gem5 works. Instead of having to dig around with the code to find
certain pieces of information, this will be a place of convenience.
For now, I am just adding information on object parameter types, but can
be expanded in the future.

Change-Id: Id51f56ab540aa7fa5ee79229aa4ab698eb5a3376