mem: Ensure data requests properly lock parent nodes
There was an issue where if the parent of a request was verified before
a request was finished being hashed, it was possible to unlock the
parent early. This is now prevented by adding locks that ensure the
parent remains in the metadata cache until every request that required
it has been satisfied.