mem: Rewrite IntegrityVerifier packet send/recv handling
This abstracts and consolidates some of the tracking and behavior when
packets are sent in and out of IntegrityVerifier.

packetLookup now is guaranteed to store a request's information until it
is sent to memory (if it doesn't need a response) or it is returned to
the CPU (if it does need a response).