mem: Prevent some packets from going out of order
There is an edge case that can cause many issues where a WritebackDirty
immediately followed by a ReadReq for the same address will cause
retrieving stale data from memory. This is due to the WritebackDirty
being processed (writes must verify before being passed to memory), but
the policy for reads is to only verify after returning from memory.
Thus, there is a possible case where the ReadReq skips past the order of
the WritebackDirty.