mem: Restructure SimpleMetadataCache into MetadataCache
This change completely overhauls the metadata cache system to have
proper set associativity. In this, functionality is also moved as needed
from SimpleMetadataCache to the AbstractMetadataCache class, and
the SimpleMetadataCache is removed in favor of MetadataCache. The
existing MetadataCache class has been replaced with the remains of
SimpleMetadataCache. The original MetadataCache class was originally
designed with a different scope than is needed now.

The LRU replacement policy is also implemented.

A new MetadataCacheEviction debug flag is also added to debug
specifically the eviction process when something fails.

The eviction process has also been rewritten to simplify the process of
finding which entries in cache are eligible to be evicted (using
getEvictable()).

This also prepares to add a second metadata cache type, but not fully
here.