stdlib: Add support for dual-memory systems in X86Board
This allows DRAM and CXL memory to be added, and choose whether one or
the other should be the "primary memory" (associated with address 0).
There are currently some issues with CXL memory being the primary
memory, which may be addressed at a later point. However, the structural
support is added, and can have both a primary and secondary memory.
"Secondary memory" is considered "the other memory that may or may not
exist and is associated with an address range starting after the primary
memory."