mem: Restructure integrity verifier to use standard cache
This reorganizes a lot of the integrity verifier component, and
simplifies a considerable portion of the behavior due to removing the
need for managing evictions and cache state. Instead, this is managed
more simply using packets and the standard cache component. This
removes support for partitioned metadata cache at this time, but is
planned to be reimplemented using the standard cache structure.